requirements.txt
google-api-python-client
pandas
python-dotenv
textblob
wordcloud
matplotlib
emoji
streamlit
plotly
